Remove $http dependency from keepalive.
Having a dependency to $http in the keepalive service causes a cyclical dependency when trying to use the idle service within an http interceptor. This will sadly remove the capability for the keepalive service to ping an external service. Users can still latch onto the 'Keepalive' event and make the $http calls themselves.